+++ packages/compiler/src/analyze.ts
@@ -23,6 +23,7 @@ import type {
   VINE_MACRO_NAMES,
   VineCompFnCtx,
   VineCompilerHooks,
+  VineDestructuredProp,
   VineFileCtx,
   VineFnPickedInfo,
   VinePropMeta,
@@ -34,6 +35,7 @@ import type {
 import {
   isArrayExpression,
   isArrayPattern,
+  isAssignmentPattern,
   isBlockStatement,
   isBooleanLiteral,
   isClassDeclaration,
@@ -45,6 +47,8 @@ import {
   isImportNamespaceSpecifier,
   isImportSpecifier,
   isObjectPattern,
+  isObjectProperty,
+  isRestElement,
   isStringLiteral,
   isTaggedTemplateExpression,
   isTemplateLiteral,
@@ -375,7 +379,51 @@ const analyzeVineProps: AnalyzeRunner = (
     // The Vine validator has guranateed there's only one formal params,
     // its type is `identifier`, and it must have an object literal type annotation.
     // Save this parameter's name as `propsAlias`
-    const propsFormalParam = formalParams[0] as Identifier
+    const propsFormalParam = formalParams[0] as (Identifier | ObjectPattern)
+    const defaultsFromDestructuredProps: Record<string, Node> = {}
+
+    // If this formal parameter has destructuring,
+    // we need to record these destructed names as `desctructedPropNames`
+    if (isObjectPattern(propsFormalParam)) {
+      for (const property of propsFormalParam.properties) {
+        if (isRestElement(property)) {
+          const restProp = property.argument as Identifier
+          vineCompFnCtx.propsDestructuredNames[restProp.name] = {
+            node: restProp,
+            isRest: true,
+          }
+        }
+        else if (isObjectProperty(property)) {
+          const propItemKey = property.key as (Identifier | StringLiteral)
+          const propItemName = (
+            isIdentifier(propItemKey)
+              ? propItemKey.name
+              : propItemKey.value
+          )
+          const data: VineDestructuredProp = {
+            node: propItemKey,
+            isRest: false,
+          }
+          if (isIdentifier(property.value) && property.value.name !== propItemName) {
+            data.alias = property.value.name
+          }
+          else if (isAssignmentPattern(property.value)) {
+            data.default = property.value.right
+            defaultsFromDestructuredProps[propItemName] = property.value.right
+          }
+
+          // Why we mark it as `SETUP_REF` instead of `PROPS_ALIASED`?
+          // Because we will actually destructure from our user-land, customized `props`
+          // which may come from `useDefaults`
+          vineCompFnCtx.bindings[propItemName] = VineBindingTypes.SETUP_REF
+          vineCompFnCtx.propsDestructuredNames[propItemName] = data
+        }
+      }
+    }
+    else {
+      vineCompFnCtx.propsAlias = propsFormalParam.name
+    }
+
     const propsTypeAnnotation = propsFormalParam.typeAnnotation
     if (!isTSTypeAnnotation(propsTypeAnnotation)) {
       return
@@ -385,14 +433,20 @@ const analyzeVineProps: AnalyzeRunner = (
     vineCompFnCtx.propsFormalParamType = typeAnnotation
 
     if (isTSTypeLiteral(typeAnnotation)) {
-      vineCompFnCtx.propsAlias = propsFormalParam.name;
       // Analyze the object literal type annotation
       // and save the props info into `vineCompFnCtx.props`
       (typeAnnotation.members as TSPropertySignature[])?.forEach((member) => {
-        if (!isIdentifier(member.key) || !member.typeAnnotation) {
+        if (
+          (!isIdentifier(member.key) && !isStringLiteral(member.key))
+          || !member.typeAnnotation
+        ) {
           return
         }
-        const propName = member.key.name
+        const propName = (
+          isIdentifier(member.key)
+            ? member.key.name
+            : member.key.value
+        )
         const propType = vineFileCtx.getAstNodeContent(member.typeAnnotation.typeAnnotation)
         const propMeta: VinePropMeta = {
           isFromMacroDefine: false,
@@ -406,7 +460,17 @@ const analyzeVineProps: AnalyzeRunner = (
           typeAnnotationRaw: propType,
         }
         vineCompFnCtx.props[propName] = propMeta
-        vineCompFnCtx.bindings[propName] = VineBindingTypes.PROPS
+
+        // If the prop is already defined as a binding at destructure,
+        // we should skip defining it as a PROPS binding.
+        vineCompFnCtx.bindings[propName] ??= VineBindingTypes.PROPS
+
+        if (defaultsFromDestructuredProps[propName]) {
+          vineCompFnCtx.props[propName].default = defaultsFromDestructuredProps[propName]
+        }
+        if (!isIdentifier(member.key)) {
+          vineCompFnCtx.props[propName].nameNeedQuoted = true
+        }
       })
     }
     else {
@@ -425,6 +489,7 @@ const analyzeVineProps: AnalyzeRunner = (
           typeChecker,
           sourceFile,
           vineCompFnCtx,
+          defaultsFromDestructuredProps,
         })
         vineCompFnCtx.props = propsInfo
       }
@@ -564,7 +629,10 @@ const analyzeVineEmits: AnalyzeRunner = (
   // If `vineEmits` is inside a variable declaration,
   // save the variable name to `vineCompFn.emitsAlias`
   if (parentVarDecl) {
-    vineCompFnCtx.emitsAlias = (parentVarDecl.id as Identifier).name
+    const emitsAlias = (parentVarDecl.id as Identifier).name
+    vineCompFnCtx.emitsAlias = emitsAlias
+    // vineEmits is treated as `setup-const` bindings #89
+    vineCompFnCtx.bindings[emitsAlias] = VineBindingTypes.SETUP_CONST
   }
 }
 
@@ -841,6 +909,7 @@ const analyzeVineModel: AnalyzeRunner = (
     }
 
     const varName = parentVarDecl.id.name
+    const typeParameter = macroCall.typeParameters?.params[0]
 
     // If the macro call has no argument,
     // - its model name is 'modelValue' as default
@@ -851,6 +920,7 @@ const analyzeVineModel: AnalyzeRunner = (
         varName,
         modelModifiersName: DEFAULT_MODEL_MODIFIERS_NAME,
         modelOptions: null,
+        typeParameter,
       }
       continue
     }
@@ -867,6 +937,7 @@ const analyzeVineModel: AnalyzeRunner = (
           varName,
           modelModifiersName: `${modelName}Modifiers`,
           modelOptions: null,
+          typeParameter,
         }
       }
       // If this argument is a object literal,
@@ -878,6 +949,7 @@ const analyzeVineModel: AnalyzeRunner = (
           varName,
           modelModifiersName: DEFAULT_MODEL_MODIFIERS_NAME,
           modelOptions: macroCall.arguments[0],
+          typeParameter,
         }
       }
     }
@@ -900,11 +972,6 @@ const analyzeVineModel: AnalyzeRunner = (
     vineCompFnCtx.bindings[modelName] = VineBindingTypes.PROPS
     // If `varName` is equal to `modelName`, it would be overrided to `setup-ref`
     vineCompFnCtx.bindings[modelDef.varName] = VineBindingTypes.SETUP_REF
-  }
-
-  // vineEmits is treated as `setup-const` bindings #89
-  if (vineCompFnCtx.emitsAlias && vineCompFnCtx.emits.length) {
-    vineCompFnCtx.bindings[vineCompFnCtx.emitsAlias] = VineBindingTypes.SETUP_CONST
   }
 }
 
@@ -1025,6 +1092,7 @@ function buildVineCompFnCtx(
     templateSource,
     templateComponentNames: new Set<string>(),
     linkedMacroCalls: [],
+    propsDestructuredNames: {},
     propsDefinitionBy: 'annotaion',
     propsAlias: 'props',
     emitsAlias: 'emits',
@@ -1083,7 +1151,7 @@ export function analyzeVine(
   vineCompilerHooks: VineCompilerHooks,
   vineFileCtx: VineFileCtx,
   vineCompFnDecls: Node[],
-) {
+): void {
   // Analyze all import statements in this file
   // and make a userImportAlias for key methods in 'vue', like 'ref', 'reactive'
   // in order to create binding records