+++ packages/compiler/tests/transform.spec.ts
@@ -250,30 +250,11 @@ export function MyComp() {
 
       typeof __VUE_HMR_RUNTIME__ !== "undefined" &&
         __VUE_HMR_RUNTIME__.createRecord(MyComp.__hmrId, MyComp);
-      export const _rerender_only = false;
-      export const _rerender_vcf_fn_name = "";
-      if (import.meta.hot) {
-        import.meta.hot.accept((mod) => {
-          if (!mod) {
-            return;
-          }
-          const { _rerender_only, _rerender_vcf_fn_name } = mod;
-          if (!_rerender_vcf_fn_name) {
-            return;
-          }
-          const component = mod[_rerender_vcf_fn_name];
-          if (_rerender_only) {
-            __VUE_HMR_RUNTIME__.rerender(component.__hmrId, component.render);
-          } else {
-            __VUE_HMR_RUNTIME__.reload(component.__hmrId, component);
-          }
-        });
-      }
       "
     `)
   })
 
-  // issue#
+  // issue#174
   it('should transform top-level await expressions', async () => {
     const { mockCompilerCtx, mockCompilerHooks } = createMockTransformCtx({
       envMode: 'development',
@@ -301,6 +282,80 @@ export async function MyComp() {
       transformed,
       { parser: 'babel-ts' },
     )
+    expect(formated).toMatchSnapshot()
+  })
+
+  // issue#192
+  it('should generate export default in correct position', async () => {
+    const { mockCompilerCtx, mockCompilerHooks } = createMockTransformCtx({
+      envMode: 'development',
+    })
+    const specContent = `
+export default function MyComp() {
+  return vine\`
+    <div>Test</div>
+  \`
+}
+    `
+    compileVineTypeScriptFile(specContent, 'testExportDefaultInCorrectPosition', { compilerHooks: mockCompilerHooks })
+    expect(mockCompilerCtx.vineCompileErrors.length).toBe(0)
+    const fileCtx = mockCompilerCtx.fileCtxMap.get('testExportDefaultInCorrectPosition')
+    const transformed = fileCtx?.fileMagicCode.toString() ?? ''
+    const formated = await format(
+      transformed,
+      { parser: 'babel-ts' },
+    )
+    expect(formated).toMatchSnapshot()
+  })
+
+  // discussion#199
+  it('should transform destructured props', async () => {
+    const { mockCompilerCtx, mockCompilerHooks } = createMockTransformCtx({
+      envMode: 'development',
+    })
+    const specContent = `
+import { ref, watch } from 'vue'
+
+export function MyComp({
+  'foo:zee': fooZee,
+  bar = 1,
+  arr,
+  ...rest
+}: {
+  'foo:zee': string,
+  bar?: number,
+  arr: boolean[],
+  other1: string,
+  other2: number,
+}) {
+
+  const test1 = ref(fooZee)
+
+  const test2 = watch(() => bar, (newVal) => {
+    console.log(newVal)
+  })
+
+  return vine\`
+    <div>{{ name }}</div>
+  \`
+}
+    `
+    compileVineTypeScriptFile(specContent, 'testTransformDestructuredProps', { compilerHooks: mockCompilerHooks })
+    expect(mockCompilerCtx.vineCompileErrors.length).toBe(0)
+
+    const fileCtx = mockCompilerCtx.fileCtxMap.get('testTransformDestructuredProps')
+    const transformed = fileCtx?.fileMagicCode.toString() ?? ''
+    const formated = await format(
+      transformed,
+      { parser: 'babel-ts' },
+    )
+
+    expect(formated).toMatch('const test1 = ref(props["foo:zee"])')
+    expect(formated).toMatch('() => props.bar')
+    expect(formated).toMatch(
+      /const __propsRestProxy = _createPropsRestProxy\(__props, \[\s*"foo:zee",\s*"bar",\s*"arr",\s*\]\);/,
+    )
+
     expect(formated).toMatchSnapshot()
   })
 })