+++ packages/tsc/src/index.ts
@@ -4,12 +4,12 @@ import type {
 } from '@vue/language-core'
 import { posix as path } from 'node:path'
 import { runTsc } from '@volar/typescript/lib/quickstart/runTsc'
+import { createVueVineLanguagePlugin, setupGlobalTypes } from '@vue-vine/language-service'
 import {
   createParsedCommandLine,
   createVueLanguagePlugin,
-  resolveVueCompilerOptions,
+  getDefaultCompilerOptions,
 } from '@vue/language-core'
-import { createVueVineLanguagePlugin, setupGlobalTypes } from '@vue-vine/language-service'
 
 const windowsPathReg = /\\/g
 
@@ -34,10 +34,11 @@ export function run() {
           }
         }
         else {
-          vueOptions = resolveVueCompilerOptions({
-            // enable strict templates by default
-            strictTemplates: true,
-          })
+          vueOptions = getDefaultCompilerOptions(
+            (void 0),
+            (void 0),
+            true,
+          )
         }
 
         languagePlugins.push(