+++ packages/compiler/CHANGELOG.md
@@ -1,5 +1,83 @@
 # @vue-vine/compiler
 
+## 0.3.10
+
+### Patch Changes
+
+- Fix language service - add missing common props for component.
+
+## 0.3.9
+
+### Patch Changes
+
+- Fix language service types issue - conflicts with implicit any
+
+## 0.3.8
+
+### Patch Changes
+
+- Move compiler HMR inject code into Vite plugin for bundler agnostic.
+
+## 0.3.7
+
+### Patch Changes
+
+- Bump version to align with vscode extension.
+
+## 0.3.6
+
+### Patch Changes
+
+- Fix HMR stability issue.
+
+## 0.3.5
+
+### Patch Changes
+
+- Fix undefined handling in HMR when ts-morph resolving.
+
+## 0.3.4
+
+### Patch Changes
+
+- Bump vue language tool v2.2.2 and fix default value in ts-morph case.
+
+## 0.3.3
+
+### Patch Changes
+
+- Re-implement the way of obtaining tsconfig and remove useless dependencies
+
+## 0.3.2
+
+### Patch Changes
+
+- Supplement some important feature support of vineModel.
+
+## 0.3.1
+
+### Patch Changes
+
+- Fix props destructure incompatible with useDefaults.
+
+## 0.3.0
+
+### Minor Changes
+
+- New feature: support props destructure, same as Vue 3.5
+
+## 0.2.9
+
+### Patch Changes
+
+- Fix ts-morph can't automatically handle tsconfig references child projects.
+
+## 0.2.7
+
+### Patch Changes
+
+- Fix incorrect position of 'export default' statement generation in Vine compiler.
+
 ## 0.2.5
 
 ### Patch Changes